+++ src/main/java/br/edu/ifba/conectairece/api/features/request/domain/dto/request/RequestPostRequestDto.java
@@ -32,5 +32,5 @@ public record RequestPostRequestDto (
     UUID profileId,
 
     @NotNull
-    Integer municipalServiceId
+    Long municipalServiceId
 ) {}
